+++ src/main/java/net/edzard/kinetic/EventType.java
@@ -1,0 +1,23 @@
+package net.edzard.kinetic;
+
+/**
+ * The event's type.
+ * @author Ed
+ */
+public enum EventType {
+	MOUSEOVER,
+	MOUSEOUT,
+	MOUSEMOVE,
+	MOUSEDOWN,
+	MOUSEUP,
+	CLICK,
+	DBLCLICK,
+	DRAGSTART,
+	DRAGMOVE,
+	DRAGEND,
+	TOUCHSTART,
+	TOUCHMOVE,
+	TOUCHEND,
+	TAP,
+	DBLTAP
+}